add new client

+++ stats/analyzer/src/main/java/ru/practicum/model/RecommendedEventProtoWrap.java
@@ -1,4 +1,4 @@
 package ru.practicum.model;
 
-public record RecommendedEventProtoWrap(int evenId, double score) {
+public record RecommendedEventProtoWrap(long evenId, double score) {
 }
